Migrate output handlers to oslo-converter-uml-ea + rename folders in latter package
