+++ packages/oslo-converter-uml-ea/index.ts
@@ -4,10 +4,10 @@ export * from './lib/EaUmlConversionService';
 export * from './lib/EaUmlConversionServiceRunner';
 export * from './lib/UriRegistry';
 export * from './lib/OutputHandlerService';
-export * from './lib/converterHandlers/AttributeConverterHandler';
-export * from './lib/converterHandlers/ElementConverterHandler';
-export * from './lib/converterHandlers/PackageConverterHandler';
-export * from './lib/converterHandlers/ConnectorConverterHandler';
+export * from './lib/converter-handlers/AttributeConverterHandler';
+export * from './lib/converter-handlers/ElementConverterHandler';
+export * from './lib/converter-handlers/PackageConverterHandler';
+export * from './lib/converter-handlers/ConnectorConverterHandler';
 export * from './lib/config/DependencyInjectionConfig';
 export * from './lib/config/EaUmlConverterConfiguration';
 export * from './lib/config/EaUmlConverterServiceIdentifier';
@@ -16,3 +16,4 @@ export * from './lib/enums/DataTypes';
 export * from './lib/enums/TagNames';
 export * from './lib/interfaces/ConverterHandler';
 export * from './lib/utils/utils';
+export * from './lib/output-handlers/JsonLdOutputHandler';
