+++ packages/oslo-converter-uml-ea/lib/OutputHandlerService.ts
@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/indent */
 import { createWriteStream } from 'fs';
 import type { IOutputHandler, QuadStore } from '@oslo-flanders/core';
-import { JsonLdOutputHandler, NQuadsOutputHandler } from '@oslo-flanders/output-handlers';
+import { JsonLdOutputHandler } from './output-handlers/JsonLdOutputHandler';
 import { inject, injectable } from 'inversify';
 import { EaUmlConverterConfiguration } from './config/EaUmlConverterConfiguration';
 import { EaUmlConverterServiceIdentifier } from './config/EaUmlConverterServiceIdentifier';
