+++ packages/oslo-converter-uml-ea/lib/converter-handlers/ConnectorConverterHandler.ts
