--- packages/oslo-output-handlers/README.md
@@ -1,8 +1,0 @@
-# OSLO Output handler
-
-> Writing the contents of an RDF store to an RDF serialization to a WritableStream (file or process.stdout)
-
-## Install
-```bash
-npm install @oslo-flanders/output-handlers
-```
