--- packages/oslo-output-handlers/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/JsonLdOutputHandler';
-export * from './lib/NQuadsOutputHandler';
