--- packages/oslo-output-handlers/package.json
@@ -1,53 +1,0 @@
-{
-  "name": "@oslo-flanders/output-handlers",
-  "version": "0.0.4-alpha.0",
-  "description": "Writing the input from an RDF store in an RDF serialization to a WritableStream",
-  "author": "Digitaal Vlaanderen <https://data.vlaanderen.be/id/organisatie/OVO002949>",
-  "homepage": "https://github.com/informatievlaanderen/OSLO-UML-Transformer/tree/main/packages/oslo-output-handlers#readme",
-  "license": "MIT",
-  "main": "index.js",
-  "directories": {
-    "lib": "lib"
-  },
-  "files": [
-    "lib"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Informatievlaanderen/OSLO-UML-Transformer.git"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.ts$": "ts-jest"
-    },
-    "testRegex": "/test/.*.test.ts$",
-    "moduleFileExtensions": [
-      "ts",
-      "js"
-    ],
-    "collectCoverage": true,
-    "coveragePathIgnorePatterns": [
-      "/node_modules/",
-      "index.js"
-    ],
-    "testEnvironment": "node"
-  },
-  "scripts": {
-    "build": "npm run build:ts",
-    "build:ts": "tsc",
-    "test:unit": "node \"../../node_modules/jest/bin/jest.js\" --group=unit"
-  },
-  "bugs": {
-    "url": "https://github.com/Informatievlaanderen/OSLO-UML-Transformer/issues"
-  },
-  "dependencies": {
-    "@oslo-flanders/core": "^0.0.4-alpha.0",
-    "n3": "^1.16.2",
-    "rdf-data-factory": "^1.1.2"
-  },
-  "devDependencies": {
-    "@rdfjs/types": "^1.1.0",
-    "@types/n3": "^1.10.4"
-  },
-  "gitHead": "f749c71908781f533e0e35f0e3394ce43a8dd7c0"
-}
